Refactor, HA compatibility adjustments
