+++ custom_components/custom_templates/templates/__init__.py
@@ -1,0 +1,1 @@
+# Templates available in custom_templates
